Start work on transformations
Part of gh-2
