+++ src/main/kotlin/geometry/Rotation.kt
@@ -1,0 +1,3 @@
+package mathx.geometry
+
+public interface Rotation : Transformation
